+++ UI/AppStyles.js
@@ -1,0 +1,34 @@
+import { StyleSheet } from "react-native";
+
+export const themeBlue = '#0b132b';
+export const themeOrange = '#FF803C';
+
+export default StyleSheet.create({
+    textBlack: {
+        color: '#000'
+    },
+    textWhite: {
+        color: '#FFF'
+    },
+    textDarkBlueBG: {
+        color: '#0b132b'
+    },
+    textOrangeBG: {
+        color: '#FF803C'
+    },
+    themeDarkBlue: {
+        backgroundColor: '#0b132b'
+    },
+    themeOrange: {
+        backgroundColor: '#FF803C'
+    },
+    header: {
+        fontFamily: 'LexendDeca_Bold'
+    },
+    body: {
+        fontFamily: 'LexendDeca'
+    },
+    title: {
+        fontFamily: 'LexendDeca_SemiBold'
+    },
+})
