+++ UI/package.json
@@ -9,12 +9,17 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-native-community/datetimepicker": "6.2.0",
     "@react-navigation/native": "^6.0.11",
     "@react-navigation/native-stack": "^6.7.0",
     "expo": "~46.0.7",
+    "expo-dev-client": "~1.2.1",
     "expo-status-bar": "~1.4.0",
     "react": "18.0.0",
     "react-native": "0.69.4",
+    "react-native-dropdown-picker": "^5.4.2",
+    "react-native-modal-datetime-picker": "^13.1.2",
+    "react-native-modal-dropdown": "^1.0.2",
     "react-native-paper": "^4.12.4",
     "react-native-safe-area-context": "4.3.1",
     "react-native-screens": "~3.15.0",
