+++ UI/react-native.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+    project: {
+        ios:{},
+        android:{}
+    },
+    assets:['./assets/fonts/'],
+}
